chore(rest-client): move normalizeOptions() function to be reusable
